refactor(shared): extract emit_terminal_tool_event into shared.terminal_events
Terminal tool UI events are shared behavior across plugins. Move the event translator into shared source so the generated plugins stay consistent.
